stock-client/requirements.txt
certifi==2020.12.5
chardet==4.0.0
comtypes==1.1.8
idna==2.10
joblib==1.0.0
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.2.0
Pillow==8.1.0
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2020.5
pywin32==300
pywinauto==0.6.8
requests==2.25.1
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
tenacity==6.3.1
threadpoolctl==2.1.0
urllib3==1.26.2
websocket-client==0.57.0
